% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{build_monitor}
\alias{build_monitor}
\title{Build the OLB monitor}
\usage{
build_monitor(ini, end, out_dir = getwd())
}
\arguments{
\item{ini}{Starting date in the YYYY-MM-DD format (\code{character}).}

\item{end}{Ending date in the YYYY-MM-DD format (\code{character}).}

\item{out_dir}{Output directory to store the monitor.}
}
\value{
A OLB's Legislative Monitor website.
}
\description{
\code{build_monitor()} summarizes legislative data and outputs a Rmarkdown report.
}
