% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olb.R
\name{build_monitor_data}
\alias{build_monitor_data}
\title{Collect all the data to create OLB's legislative monitor}
\usage{
build_monitor_data()
}
\description{
\code{build_monitor_data()} create a dataset with the required data to update the OLB's monitor website.
}
