% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputies.R
\name{get_deputies_bios}
\alias{get_deputies_bios}
\title{Get a deputy's bio}
\usage{
get_deputies_bios(...)
}
\arguments{
\item{...}{query parameters passed to the get_deputies_list call (inside the function).}
}
\value{
A \code{tibble}.
}
\description{
\code{get_deputies_bios()} extracts the biography of all seating deputies.
}
