% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputies.R
\name{get_deputies_list}
\alias{get_deputies_list}
\title{Get a list of all deputies}
\usage{
get_deputies_list(...)
}
\arguments{
\item{...}{query parameters.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_deputies_list()} extracts a list of national deputies in Bazil.
}
