% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deputies.R
\name{get_deputy_bio}
\alias{get_deputy_bio}
\title{Get a deputy's bio}
\usage{
get_deputy_bio(dep_id)
}
\arguments{
\item{dep_id}{A deputy's ID.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_deputy_bio()} extracts a deputy's biography.
}
