% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olb.R
\name{get_monitor_data}
\alias{get_monitor_data}
\title{Collect data to update the OLB's legislative monitor}
\usage{
get_monitor_data(dataInicio, dataFim, codTema)
}
\arguments{
\item{dataInicio}{Start date (in YYYY-MM-DD format, as \code{character}).}

\item{dataFim}{End date (in YYYY-MM-DD format, as \code{character}).}

\item{codTema}{Topic code (use \code{\link{list_topics}} to obtain a list of all valid topic codes).}
}
\value{
A \code{tibble} with some \code{nested} columns.
}
\description{
\code{get_monitor_data()} gathers a dataset with all the information required to produce
OLB's llegislative monitor.
}
