% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{get_proposal}
\alias{get_proposal}
\title{Get a legislative proposal}
\usage{
get_proposal(id)
}
\arguments{
\item{id}{Proposal id.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_proposal()} extracts detailed information about a given legislative proposal.
}
