% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{get_proposal_author}
\alias{get_proposal_author}
\title{Get the authors of a certain proposal}
\usage{
get_proposal_author(id)
}
\arguments{
\item{id}{Main proposal id.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_proposal_author()} extracts a list of authors of a given legislative proposal.
}
