% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{get_proposal_history}
\alias{get_proposal_history}
\title{Get the history of a certain proposal}
\usage{
get_proposal_history(id)
}
\arguments{
\item{id}{Main proposal id.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_proposal_history()} extracts a list of the situations of a given legislative proposal.
}
