% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{get_proposal_list}
\alias{get_proposal_list}
\title{Get a list of proposals}
\usage{
get_proposal_list(...)
}
\arguments{
\item{...}{query parameters.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_proposal_list()} extracts a list of proposals introduced in the Brazilian Chamber of Deputies.
}
