% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{get_proposal_topic}
\alias{get_proposal_topic}
\title{Get the topics of a certain proposal}
\usage{
get_proposal_topic(id)
}
\arguments{
\item{id}{Main proposal id.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_proposal_topic()} a list of topics related to a given legislative proposal.
}
