% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{get_proposals_olb}
\alias{get_proposals_olb}
\title{Get a list of proposals of a given type}
\usage{
get_proposals_olb(...)
}
\arguments{
\item{...}{query parameters.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_proposals_olb()} extracts a list of PEC, PL, MPV, and PLP introduced in the Brazilian Chamber of Deputies.
}
