% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{get_related_proposals}
\alias{get_related_proposals}
\title{Get a list of proposals related to a certain one}
\usage{
get_related_proposals(id)
}
\arguments{
\item{id}{Main proposal id.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_related_proposals()} a list of proposals related to a given legislative proposal.
}
