% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech.R
\name{get_speech}
\alias{get_speech}
\title{Get legislative speech}
\usage{
get_speech(dep_id, ...)
}
\arguments{
\item{dep_id}{A deputy's ID.}

\item{...}{query parameters.}
}
\value{
A \code{tibble}.
}
\description{
\code{get_speech()} extracts speech data by deputy.
}
