% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{list_proposal_situations}
\alias{list_proposal_situations}
\title{List all the possible situations of legislative bills}
\usage{
list_proposal_situations()
}
\value{
A \code{tibble}.
}
\description{
\code{list_ropics()} returns a \code{tibble} with a list of situations used by the Brazilian Chamber
of Deputies to track legislative bills.
}
