% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{list_topics}
\alias{list_topics}
\title{List all the topics of legislative bills}
\usage{
list_topics()
}
\value{
A \code{tibble}.
}
\description{
\code{list_ropics()} returns a \code{tibble} with a list of topics used by the Brazilian Chamber
of Deputies to classify legislative bills.
}
