% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista_temas.R
\name{lista_temas}
\alias{lista_temas}
\title{List all the themes of legislative bills}
\usage{
lista_temas()
}
\value{
A \code{data.frame}.
}
\description{
\code{lista_temas()} returns a \code{data.frame} with a list of themes used by the Brazilian Chamber
of Deputies to classify legislative bills.
}
