% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_autores.R
\name{pega_autores}
\alias{pega_autores}
\title{Get authors of legislative bills}
\usage{
pega_autores(year)
}
\arguments{
\item{year}{Year in YYYY format (\code{character} or \code{numeric}).}
}
\value{
A \code{data.frame}.
}
\description{
\code{pega_autores()} is a convenience function used to download data about the authors of the legislative
bills introduced in the Brazilian Chamber of Deputies' website.
}
\examples{
\dontrun{
# Get information for all bills in 2010
pega_autores(2010)
}

}
