% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_proposicoes.R
\name{pega_props}
\alias{pega_props}
\title{Extract legislative bills from the Brazilian Chamber of Deputies' website}
\usage{
pega_props(year)
}
\arguments{
\item{year}{Year in AA format (\code{character} or \code{numeric}).}
}
\value{
A \code{data.frame} containing information for each bill proposed in a given year.
}
\description{
\code{pega_propos()} is a convenience function used to download data on legislative bills.
}
\examples{
\dontrun{
# Get information on all bills in 2010
pega_propos(2010)
}

}
