% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_tema.R
\name{pega_tema}
\alias{pega_tema}
\title{Extract legislative bills' classification}
\usage{
pega_tema(year)
}
\arguments{
\item{year}{Year in YYYY format (\code{character} or \code{numeric}).}
}
\value{
A \code{data.frame}.
}
\description{
\code{pega_propos()} is a convenience function used to download data on legislative bills's
classification from the Brazilian Chamber of Deputies' website.
}
\examples{
\dontrun{
# Get information on all bills in 2010
pega_tema(2010)
}

}
