% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olb.R
\name{proposals_track_report}
\alias{proposals_track_report}
\title{Collect data to produce legislative reports by topic}
\usage{
proposals_track_report(dataInicio, dataFim, codTema)
}
\arguments{
\item{dataInicio}{Start date (in YYYY-MM-DD format, as \code{character}).}

\item{dataFim}{End date (in YYYY-MM-DD format, as \code{character}).}

\item{codTema}{Topic code (use \code{\link{list_topics}} to obtain a list of all valid topic codes).}
}
\value{
A \code{tibble} with some \code{nested} columns.
}
\description{
\code{proposals_track_report()} gathers a dataset with all the information required to produce
OLB's legislative reports by topic.
}
