% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposicoes.R
\name{proposicoes}
\alias{proposicoes}
\title{Arrange data on legislative bills introduced in a given period}
\usage{
proposicoes(ini, end)
}
\arguments{
\item{ini}{Starting date in the YYYY-MM-DD format (\code{character}).}

\item{end}{Ending date in the YYYY-MM-DD format (\code{character}).}
}
\value{
A \code{data.frame} containing selected information for each bill introduced in a given year.
}
\description{
\code{proposicoes()} cleans data on legislative bills introduced in a given period.
}
