% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temas.R
\name{tema_olb}
\alias{tema_olb}
\title{Tema do ggplot2}
\usage{
tema_olb(posicao = "none", base_size = 14, base_family = "worksans")
}
\arguments{
\item{posicao}{Lugar da legenda}

\item{base_size}{Tamanho da fonte}

\item{base_family}{Fonte}
}
\description{
Tema do ggplot2
}
