% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_tramitacao.R
\name{tramitacao}
\alias{tramitacao}
\title{Extract information about the legislative activity of a given bill}
\usage{
tramitacao(id_prop)
}
\arguments{
\item{id_prop}{Bill ID as used by the Brazilian Chamber of Deputies (\code{character} or \code{numeric}).}
}
\value{
A \code{data.frame} containing information for a given bill.
}
\description{
\code{pega_tramitacao()} extracts information about the activity of a given legislative bill introduced in the
Brazilian Chamber of Deputies.
}
\examples{
\dontrun{
# Get information on a given bill
tramitacao(id_prop = 123)
}

}
