% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_relatorio.R
\name{gera_relatorio}
\alias{gera_relatorio}
\title{Produce a report of legislative activity in a given period}
\usage{
gera_relatorio(ini, end)
}
\arguments{
\item{ini}{Starting date in the YYYY-MM-DD format (\code{character}).}

\item{end}{Ending date in the YYYY-MM-DD format (\code{character}).}
}
\value{
A report and its corresponding .Rmd file.
}
\description{
\code{gera_relatorio()} summarizes legislative data and outputs a Rmarkdown report.
}
